# lite: memStoreProvider GetHeightBinarySearch method + fix ValKeys.signHeaders