@melekes 2018-01-08
https://github.com/tendermint/tendermint/pull/1043#discussion_r160029186