@melekes 2018-01-08
```
lite/memprovider.go:92:8:warning: "heighest" is a misspelling of "heights" (misspell)
lite/performance_test.go:201:5:warning: should omit comparison to bool constant, can be simplified to !algo (S1002) (gosimple)
```