@ebuchman 2018-01-24
Can we get a test that shows they return the same result for various queries on variously filled stores ?