@odeke-em 2018-01-25
Alright, @melekes @ebuchman, I've added the requested tests with commit https://github.com/tendermint/tendermint/pull/1043/commits/da08b7b2eca3ac81848ca2ffd4526a89b303f023, PTAL.