@odeke-em 2018-01-25
@adrianbrink what's the linting error?