@zramsay 2018-01-25
https://circleci.com/gh/tendermint/tendermint/3003?utm_campaign=vcs-integration-link&utm_medium=referral&utm_source=github-build-link

`lite/performance_test.go:201:5:warning: should omit comparison to bool constant, can be simplified to !algo (S1002) (gosimple)`