@odeke-em 2018-01-29
> lite/performance_test.go:201:5:warning: should omit comparison to bool constant, can be simplified to !algo (S1002) (gosimple)

Thanks for the notice @zramsay. Hmm, that lint warning and suggestion lose the point of readability though. I presume that we could change that to an int, but still I don't think we should be changing up readability to match a linter.