@zramsay 2018-01-29
yeah that's why i added `nolint` instead of actually linting it