@odeke-em 2018-02-01
> yeah that's why i added nolint instead of actually linting it
@zramsay oh I see, thank you!

> Seems fine on the surface, but needs to be properly integrated - doesn't make sense to have the two functions exposed, since we're implementing an interface here, and no user should be responsible for choosing.
Great point @ebuchman. Yes in deed, I posted up that perhaps we need an evaluation of when to switch to using the getHeightBinarySearch perhaps after hitting the point when linear search is no longer fast, perhaps derived from the benchmarks. I didn't yet know though if the PR would even be accepted, and since it has been, let me add that. Thank you for the ping.

> As is, the GetByHeightBinarySearch isn't really useful, as it's not exposed on the interface and not implemented for the other provider implementations, so I think we should address that before merging.
100% in deed, that's the last step now that it's been accepted

as I had mentioned in https://github.com/tendermint/tendermint/pull/1043#issue-285188242
> If need be we can make a hybrid algorithm that switches between the
linear and binary search depending on the number of items.
This is reminiscent of Python's TimSort algorithm.