@odeke-em 2018-02-01
@ebuchman I've pushed a PR updating with the requested changes, PTAL and thank you for the review!